+++ src/kernel/impl/interrupts/idt.c
@@ -1,7 +1,12 @@
 #include "stddef.h"
 #include "stdint.h"
 #include "idt.h"
+#include "util.h"
 
+#define PIC1_CMD        0x20
+#define PIC1_DATA       0x21
+#define PIC2_CMD        0xA0
+#define PIC2_DATA       0xA1
 #define IDT_DESCRIPTORS 256
 
 extern void *isr_stub_table[];   // generate array of asm stubs
@@ -22,6 +27,28 @@ static void setIdtEntry(int intex, void *isr, uint8_t flags, uint8_t ist) {
 }
 
 void idtInit(void) {
+
+    // === PIC ===
+    // Start initialization sequence (cascade mode, expect ICW4)
+    outb(PIC1_CMD, 0x11);
+    outb(PIC2_CMD, 0x11);
+
+    // Remap offsets: master to 0x20 (32), slave to 0x28 (40)
+    outb(PIC1_DATA, 0x20);
+    outb(PIC2_DATA, 0x28);
+
+    // Setup cascading
+    outb(PIC1_DATA, 0x04); // Tell master that slave is at IRQ2
+    outb(PIC2_DATA, 0x02); // Tell slave its cascade identity
+
+    // Environment info
+    outb(PIC1_DATA, 0x1);
+    outb(PIC2_DATA, 0x1);
+
+    // Mask all IRQs except IRQ0 (timer) (for now since nothing else can't be handelled yet)
+    outb(PIC1_DATA, 0xFE);
+    outb(PIC2_DATA, 0xFF);
+
     // Fill exception handlers (0–31)
     for (int i = 0; i < 32; i++) {
         setIdtEntry(i, isr_stub_table[i], 0x8E, 0); // ring0, interrupt gate
