--- src/kernel/intf/pic.h
@@ -1,3 +1,0 @@
-#pragma once
-
-void pic_init(void);
